Change ChainEpoch and TokenAmount types (#309)
* Migrate tokenamount to an alias

* Migrate ChainEpoch to an alias

* Clean up references and unused test imports

+++ node/forest_libp2p/tests/decode_test.rs
@@ -44,11 +44,11 @@ fn convert_single_tipset_bundle() {
 #[test]
 fn tipset_bundle_to_full_tipset() {
     let h0 = BlockHeader::builder()
-        .weight(BigUint::from(1 as u32))
+        .weight(BigUint::from(1u32))
         .build()
         .unwrap();
     let h1 = BlockHeader::builder()
-        .weight(BigUint::from(2 as u32))
+        .weight(BigUint::from(2u32))
         .build()
         .unwrap();
     let ua = UnsignedMessage::builder()
